requirements.txt
numpy==2.3.1
pygame==2.6.1
pyglm==2.8.2
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
